requirements.txt
# Core ML
onnxruntime-gpu==1.19.2
faiss-cpu==1.8.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# GUI
PySide6>=6.6.0

# Utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
Pillow>=10.0.0

# Development
pytest>=7.4.0
black>=23.0.0